+++ addOns/portscan/CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 ### Changed
+- Use the Network add-on to obtain the outgoing proxy.
 - Maintenance changes.
 - Update minimum ZAP version to 2.11.1.